created lxa global Runner
